third_party/blink/web_tests/platform/generic/http/tests/navigation/anchor-frames-expected.txt



--------
Frame: 'main'
--------
Tests that loading a frame with a URL that contains a fragment pointed at a named anchor actually scrolls to that anchor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.documentElement.offsetHeight > document.documentElement.clientHeight is true
PASS document.scrollingElement.scrollTop > 0 is true
PASS document.scrollingElement.scrollTop + document.body.clientHeight > 2000 is true
PASS successfullyParsed is true

TEST COMPLETE

This is an anchor point named "anchor1.

--------
Frame: 'footer'
--------
